+++ otrs/operations.py
@@ -8,7 +8,7 @@ import requests, json
 import pyotrs
 from connectors.core.connector import get_logger, ConnectorError
 
-logger = get_logger('otrs')
+logger = get_logger('otrs_dev')
 
 
 def _create_client(config, **kwargs):
@@ -25,6 +25,10 @@ def _create_client(config, **kwargs):
 def create_ticket(config, params):
     try:
         client = _create_client(config)
+    except ConnectorError as e:
+        raise e
+
+    try:
         title = params.get("title")
         queue = params.get("queue")
         state = params.get("state")
@@ -32,30 +36,29 @@ def create_ticket(config, params):
         customer = params.get("customer")
         article_subject = params.get("article_subject")
         article_body = params.get("article_body")
-        update_params = {}
-
-        if params.get("dynamicField"):
-            name = params.get("dynamicField")['Name']
-            value = params.get("dynamicField")['Value']
-            df = pyotrs.DynamicField(name, value)
-            update_params.update(dynamic_fields=[df])
-
-        ticket = pyotrs.Ticket.create_basic(Title=title, Queue=queue, State=state, Priority=priority,
-                                            CustomerUser=customer)
+        ticketType = params.get('newTicketType')
+        logger.error('ticketType:{0}'.format(str(ticketType)))
+        ticket = pyotrs.Ticket.create_basic(Title=title, Queue=queue, State=state, Priority=priority, CustomerUser=customer, Type=ticketType)
         if not article_subject:
             article_subject = title
         ticket_article = pyotrs.Article({"Subject": article_subject, "Body": article_body})
-        update_params.update({"article": ticket_article})
-        ticket_info = client.ticket_create(ticket, **update_params)
+        ticket_info = client.ticket_create(ticket, ticket_article)
 
         return {"ticket_metadata": ticket_info}
-    except Exception as Err:
-        raise ConnectorError(Err)
+    except Exception as e:
+        error_message = "Error creating OTRS ticket. Error message as follows:\n{0}".format(str(e))
+        logger.exception(error_message)
+        raise ConnectorError(error_message)
 
 
 def update_ticket(config, params):
+    logger.error('PARAMS {0}'.format(params))
     try:
         client = _create_client(config)
+    except ConnectorError as e:
+        raise e
+
+    try:
         ticket_id = params["ticket_id"]
         update_params = {}
         if params.get("title"):
@@ -68,49 +71,87 @@ def update_ticket(config, params):
             update_params.update({"Priority": params["priority"]})
         if params.get("customer"):
             update_params.update({"CustomerUser": params["customer"]})
-        if params.get("dynamicField"):
-            name = params.get("dynamicField")['Name']
-            value = params.get("dynamicField")['Value']
-            df = pyotrs.DynamicField(name, value)
-            update_params.update(dynamic_fields=[df])
-        if params.get("article_subject") and params.get("article_body"):
-            new_article = pyotrs.Article({"Subject": params.get("article_subject"), "Body": params.get("article_body")})
-            update_params.update({"article": new_article})
+        if params.get("oTRSArticle"):
+            newArticle = pyotrs.Article(params.get("oTRSArticle"))
+            update_params.update(article=newArticle)
+            
+        if params["dynamicField"]:
+            logger.error('PARAMS {0}'.format(str(params["dynamicField"])))
+            dynamicFieldsParameter = params["dynamicField"]
+            dFields = []
+            for item in dynamicFieldsParameter:
+              logger.error('ITEM {0}'.format(str(item)))
+              name = item['Name']
+              value = item['Value']
+              dFields.append(pyotrs.DynamicField(str(name), str(value)))
+            logger.error('dFields {0}'.format(str(dFields)))
+            update_params.update(dynamic_fields=dFields)
+            
 
-        ticket_info = client.ticket_update(ticket_id, **update_params)
+        ticket_info = client.ticket_update(ticket_id, **update_params )
 
         return {"ticket_metadata": ticket_info}
-    except Exception as Err:
-        raise ConnectorError(Err)
+    except Exception as e:
+        error_message = "Error updating ticket {0}. Error message as follows:\n{1}".format(ticket_id, str(e))
+        logger.exception(error_message)
+        raise ConnectorError(error_message)
 
 
 def get_ticket(config, params):
     try:
         client = _create_client(config)
+    except ConnectorError as e:
+        raise e
+
+    try:
         ticket_id = params["ticket_id"]
         ticket = client.ticket_get_by_id(ticket_id, articles=True, attachments=False, dynamic_fields=True)
 
         return {"ticket": ticket.to_dct()}
-    except Exception as Err:
-        raise ConnectorError(Err)
+    except Exception as e:
+        error_message = "Error fetching OTRS ticket {0}. Error message as follows:\n{1}".format(ticket_id, str(e))
+        logger.exception(error_message)
+        raise ConnectorError(error_message)
 
 
 def search_tickets(config, params):
     try:
         client = _create_client(config)
-        state = params.get('state')
-        title = params.get('title')
-        ticketType = params.get('tickettype')
+    except ConnectorError as e:
+        raise e
 
-        matching_tickets = client.ticket_search(StateType=state, Types=ticketType)
-
+    try:
+        states = params.get('state')
+        title= params.get('title')
+        ticketType=params.get('tickettype')
+        logger.error('state {0}'.format(str(states)))
+        logger.error('Ticket Type {0}'.format(str(ticketType)))
+        timeSpanMinutes=str(params.get('timeSpanMinutes'))
+        
+        if states and ticketType:
+          matching_tickets = client.ticket_search(States=states, Types=ticketType)
+          return {"count": len(matching_tickets),"matches": matching_tickets}
+		  
+        if states:
+          matching_tickets = client.ticket_search(States=states, Types=ticketType)
+          return {"count": len(matching_tickets),"matches": matching_tickets}
+        
+        if timeSpanMinutes and title:
+          matching_tickets = client.ticket_search(Title=title,TicketChangeTimeNewerMinutes=timeSpanMinutes,States=states,Types=ticketType)
+          return {"count": len(matching_tickets),"matches": matching_tickets}
+                
         if title:
-            matching_tickets = client.ticket_search(Title=title, StateType=state, Types=ticketType)
-
-        return {"count": len(matching_tickets),
-                "matches": matching_tickets}
-    except Exception as Err:
-        raise ConnectorError(Err)
+          matching_tickets = client.ticket_search(Title=title,States=states,Types=ticketType)
+          return {"count": len(matching_tickets),"matches": matching_tickets}
+        
+        if timeSpanMinutes:
+          matching_tickets = client.ticket_search(TicketChangeTimeNewerMinutes=timeSpanMinutes,States=states,Types=ticketType)
+          return {"count": len(matching_tickets), "matches": matching_tickets}
+        
+    except Exception as e:
+        error_message = "Error performing ticket search. Error message as follows:\n{0}".format(str(e))
+        logger.exception(error_message)
+        raise ConnectorError(error_message)
 
 
 def check_health(config):
@@ -118,8 +159,8 @@ def check_health(config):
     try:
         client = _create_client(config)
         return True
-    except Exception as Err:
-        error_message = "Error in health check. Error message as follows:{0}".format(str(Err))
+    except Exception as e:
+        error_message = "Error in health check. Error message as follows:{0}".format(str(e))
         logger.exception(error_message)
         raise ConnectorError(error_message)
 
@@ -128,5 +169,6 @@ operations = {
     'create_ticket': create_ticket,
     'update_ticket': update_ticket,
     'get_ticket': get_ticket,
-    'search_tickets': search_tickets
+    'search_tickets': search_tickets,
+    'check_health': check_health
 }